// client/src/app/models/printGatePassContainer.ts
export interface IGatePass {
  containerNumber: string
  transactionDate: string
  deliveryType: string
  consignee: string
  invoiceNumber: string
  grossWeight: string
  containerLength: string
  containerFreightStatus: string
  containerCategory: string
  lineId: string
  containerIsoCode: string
  berth: string
  finalValidityDate: string
  customerContract: string;
}